mod: code reorganized
