+++ tools/test_delete.h
